+++ package.json
@@ -13,6 +13,7 @@
     "react": "^18.1.0",
     "react-bootstrap": "^2.5.0",
     "react-dom": "^18.1.0",
+    "react-icons": "^4.6.0",
     "react-router-dom": "^6.3.0",
     "sass": "^1.52.3"
   },
